requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
openai>=1.3.0
python-multipart>=0.0.6
